Issue-1776-2: removed all tests from PipelineRunManagerTest (except 2) to right places.
