+++ api/src/test/java/com/epam/pipeline/manager/pipeline/PipelineRunManagerLaunchTest.java
@@ -18,6 +18,7 @@ package com.epam.pipeline.manager.pipeline;
 
 import com.epam.pipeline.common.MessageHelper;
 import com.epam.pipeline.dao.pipeline.PipelineRunDao;
+import com.epam.pipeline.entity.BaseEntity;
 import com.epam.pipeline.entity.cluster.InstancePrice;
 import com.epam.pipeline.entity.cluster.PriceType;
 import com.epam.pipeline.entity.configuration.PipeConfValueVO;
@@ -31,36 +32,76 @@ import com.epam.pipeline.manager.cluster.InstanceOfferManager;
 import com.epam.pipeline.manager.datastorage.DataStorageManager;
 import com.epam.pipeline.manager.execution.PipelineLauncher;
 import com.epam.pipeline.manager.preference.AbstractSystemPreference;
+import com.epam.pipeline.entity.pipeline.TaskStatus;
+import com.epam.pipeline.entity.pipeline.Tool;
+import com.epam.pipeline.entity.pipeline.run.RunStatus;
+import com.epam.pipeline.entity.region.AwsRegion;
+import com.epam.pipeline.manager.cluster.InstanceOfferManager;
+import com.epam.pipeline.manager.datastorage.DataStorageManager;
+import com.epam.pipeline.manager.docker.ToolVersionManager;
+import com.epam.pipeline.manager.execution.PipelineLauncher;
 import com.epam.pipeline.manager.preference.PreferenceManager;
 import com.epam.pipeline.manager.region.CloudRegionManager;
 import com.epam.pipeline.manager.security.AuthManager;
 import com.epam.pipeline.manager.security.CheckPermissionHelper;
 import com.epam.pipeline.manager.security.run.RunPermissionManager;
+import com.epam.pipeline.util.TestUtils;
+import io.reactivex.subjects.BehaviorSubject;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
+import java.time.LocalDateTime;
 import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+import static com.epam.pipeline.common.MessageConstants.ERROR_INSTANCE_TYPE_IS_NOT_ALLOWED;
+import static com.epam.pipeline.common.MessageConstants.ERROR_PRICE_TYPE_IS_NOT_ALLOWED;
+import static com.epam.pipeline.common.MessageConstants.ERROR_RUN_CLOUD_REGION_NOT_ALLOWED;
+import static com.epam.pipeline.common.MessageConstants.ERROR_TOOL_CLOUD_REGION_NOT_ALLOWED;
 
 import static com.epam.pipeline.entity.contextual.ContextualPreferenceLevel.TOOL;
 import static com.epam.pipeline.manager.preference.SystemPreferences.CLUSTER_DOCKER_EXTRA_MULTI;
 import static com.epam.pipeline.manager.preference.SystemPreferences.CLUSTER_INSTANCE_HDD_EXTRA_MULTI;
 import static com.epam.pipeline.manager.preference.SystemPreferences.CLUSTER_SPOT;
 import static com.epam.pipeline.manager.preference.SystemPreferences.COMMIT_TIMEOUT;
+
 import static com.epam.pipeline.test.creator.configuration.ConfigurationCreatorUtils.getPipelineConfiguration;
 import static com.epam.pipeline.test.creator.docker.DockerCreatorUtils.getTool;
+
+import static com.epam.pipeline.test.creator.CommonCreatorConstants.ID;
+import static com.epam.pipeline.test.creator.CommonCreatorConstants.ID_2;
+import static com.epam.pipeline.test.creator.CommonCreatorConstants.ID_3;
+import static com.epam.pipeline.test.creator.configuration.ConfigurationCreatorUtils.getPipelineConfiguration;
+import static com.epam.pipeline.test.creator.docker.DockerCreatorUtils.getTool;
+import static com.epam.pipeline.test.creator.pipeline.PipelineCreatorUtils.getPipelineRun;
+
 import static com.epam.pipeline.test.creator.pipeline.PipelineCreatorUtils.getPipelineRunWithInstance;
 import static com.epam.pipeline.test.creator.region.RegionCreatorUtils.getDefaultAwsRegion;
 import static com.epam.pipeline.test.creator.region.RegionCreatorUtils.getNonDefaultAwsRegion;
 import static com.epam.pipeline.util.CustomAssertions.assertThrows;
 import static java.lang.Integer.parseInt;
 import static org.hamcrest.core.Is.is;
+
 import static org.junit.Assert.assertThat;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyBoolean;
 import static org.mockito.Matchers.anyInt;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThat;
+import static org.mapstruct.ap.internal.util.Collections.newArrayList;
+import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.anyBoolean;
+import static org.mockito.Matchers.anyInt;
+import static org.mockito.Matchers.anyListOf;
+
 import static org.mockito.Matchers.anyLong;
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Matchers.eq;
@@ -73,10 +114,15 @@ public class PipelineRunManagerLaunchTest {
     private static final float PRICE_PER_HOUR = 12F;
     private static final float COMPUTE_PRICE_PER_HOUR = 11F;
     private static final float DISK_PRICE_PER_HOUR = 1F;
+
     private static final long REGION_ID = 1L;
     private static final long NON_ALLOWED_REGION_ID = 2L;
     private static final long NON_DEFAULT_REGION_ID = 3L;
     private static final String INSTANCE_TYPE = "m5.large";
+
+    private static final String NOT_ALLOWED = "not allowed";
+    private static final String NO_PERMISSIONS = "no permissions";
+
     private static final long PARENT_RUN_ID = 5L;
     private static final String INSTANCE_DISK = "1";
     private static final String PARENT_RUN_ID_PARAMETER = "parent-id";
@@ -85,7 +131,13 @@ public class PipelineRunManagerLaunchTest {
     private static final String ON_DEMAND = PriceType.ON_DEMAND.getLiteral();
     private static final String DEFAULT_COMMAND = "sleep";
     private static final String TEST_USER = "user";
+
     private static final String IMAGE = "testImage";
+
+    private static final LocalDateTime TEST_PERIOD = LocalDateTime.of(2019, 4, 2, 0, 0);
+    private static final int HOURS_12 = 12;
+    private static final int HOURS_18 = 18;
+
 
     @InjectMocks
     private final PipelineRunManager pipelineRunManager = new PipelineRunManager();
@@ -118,6 +170,9 @@ public class PipelineRunManagerLaunchTest {
     private CloudRegionManager cloudRegionManager;
 
     @Mock
+    private ToolVersionManager toolVersionManager;
+
+    @Mock
     private CheckPermissionHelper permissionHelper;
 
     @Mock
@@ -126,18 +181,24 @@ public class PipelineRunManagerLaunchTest {
     @Mock
     private PreferenceManager preferenceManager;
 
+    @Mock
+    private RunStatusManager runStatusManager;
+
     private final Tool tool = getTool(IMAGE, DEFAULT_COMMAND);
-    private final AwsRegion defaultAwsRegion = getDefaultAwsRegion(REGION_ID);
-    private final AwsRegion nonAllowedAwsRegion = getNonDefaultAwsRegion(NON_ALLOWED_REGION_ID);
-    private final AwsRegion notDefaultAwsRegion = getNonDefaultAwsRegion(NON_DEFAULT_REGION_ID);
+    private final AwsRegion defaultAwsRegion = getDefaultAwsRegion(ID);
+    private final AwsRegion nonAllowedAwsRegion = getNonDefaultAwsRegion(ID_2);
+    private final AwsRegion notDefaultAwsRegion = getNonDefaultAwsRegion(ID_3);
+
     private final PipelineConfiguration configuration = getPipelineConfiguration(
             IMAGE, INSTANCE_DISK, true, defaultAwsRegion.getId());
     private final PipelineConfiguration configurationWithoutRegion = getPipelineConfiguration(IMAGE, INSTANCE_DISK);
     private final InstancePrice price = new InstancePrice(configuration.getInstanceType(),
             parseInt(configuration.getInstanceDisk()), PRICE_PER_HOUR, COMPUTE_PRICE_PER_HOUR, DISK_PRICE_PER_HOUR);
+
     private final PipelineRun parentRun = getPipelineRunWithInstance(PARENT_RUN_ID, TEST_USER, NON_DEFAULT_REGION_ID);
     private final ContextualPreferenceExternalResource resource = new ContextualPreferenceExternalResource(
             TOOL, tool.getId().toString());
+
 
     @Before
     public void setUp() throws Exception {
@@ -172,6 +233,15 @@ public class PipelineRunManagerLaunchTest {
         launchTool(configuration, INSTANCE_TYPE);
 
         verify(instanceOfferManager).isPriceTypeAllowed(eq(SPOT), eq(resource), eq(false));
+    }
+
+    @Test
+    public void launchPipelineShouldValidateToolInstanceTypeAndPriceType() {
+        launchTool(configuration, INSTANCE_TYPE);
+
+        verify(toolManager).loadByNameOrId(eq(IMAGE));
+        verify(instanceOfferManager).isToolInstanceAllowed(eq(INSTANCE_TYPE), eq(getResource()), eq(ID), eq(true));
+        verify(instanceOfferManager).isPriceTypeAllowed(eq(SPOT), eq(getResource()), eq(false));
     }
 
     @Test
@@ -368,6 +438,105 @@ public class PipelineRunManagerLaunchTest {
         doReturn(tool).when(toolManager).resolveSymlinks(eq(IMAGE));
     }
 
+    @Test
+    public void shouldLoadRunsActivityStats() {
+        final PipelineRun pipelineRun = getPipelineRun(ID, TEST_USER);
+        final PipelineRun pipelineRun2 = getPipelineRun(ID_2, TEST_USER);
+        doReturn(pipelineRun).when(pipelineRunDao).loadPipelineRun(anyLong());
+        doReturn(newArrayList(pipelineRun, pipelineRun2)).when(pipelineRunDao)
+                .loadPipelineRunsActiveInPeriod(eq(TEST_PERIOD), eq(TEST_PERIOD.plusHours(HOURS_18)));
+        doReturn(getStatusMap()).when(runStatusManager).loadRunStatus(anyListOf(Long.class));
+
+        launchPipelineRun(TEST_PERIOD, TEST_PERIOD.plusHours(HOURS_12));
+        launchPipelineRun(TEST_PERIOD.plusHours(HOURS_18), null);
+
+        final Map<Long, PipelineRun> stats =
+                pipelineRunManager.loadRunsActivityStats(TEST_PERIOD, TEST_PERIOD.plusHours(HOURS_18)).stream()
+                        .collect(Collectors.toMap(BaseEntity::getId,
+                                Function.identity()));
+
+        assertEquals(2, stats.size());
+        assertEquals(2, stats.get(ID).getRunStatuses().size());
+        assertEquals(1, stats.get(ID_2).getRunStatuses().size());
+    }
+
+    private void launchPipelineRun(final LocalDateTime startDate, final LocalDateTime stopDate) {
+        final PipelineRun pipelineRun = launchPipeline(configuration, INSTANCE_TYPE);
+        pipelineRun.setStartDate(TestUtils.convertLocalDateTimeToDate(startDate));
+        saveStatusForRun(pipelineRun.getId(), TaskStatus.RUNNING, startDate);
+        if (stopDate != null) {
+            pipelineRun.setStatus(TaskStatus.STOPPED);
+            pipelineRun.setEndDate(TestUtils.convertLocalDateTimeToDate(stopDate));
+            saveStatusForRun(pipelineRun.getId(), TaskStatus.STOPPED, stopDate);
+        } else {
+            pipelineRun.setEndDate(null);
+        }
+        pipelineRunManager.updateRunInfo(pipelineRun);
+    }
+
+    private Map<Long, List<RunStatus>> getStatusMap() {
+        final RunStatus status1 = new RunStatus(ID, TaskStatus.RUNNING, null, TEST_PERIOD);
+        final RunStatus status2 = new RunStatus(ID, TaskStatus.STOPPED, null, TEST_PERIOD.plusHours(HOURS_18));
+        final List<RunStatus> list1 = newArrayList(status1, status2);
+        final RunStatus status3 = new RunStatus(ID_2, TaskStatus.RUNNING, null, TEST_PERIOD.plusHours(HOURS_18));
+        final List<RunStatus> list2 = newArrayList(status3);
+        final Map<Long, List<RunStatus>> map = new HashMap<>();
+        map.put(ID, list1);
+        map.put(ID_2, list2);
+        return map;
+    }
+
+    private void saveStatusForRun(final Long runId, final TaskStatus status, final LocalDateTime timePoint) {
+        runStatusManager.saveStatus(new RunStatus(runId, status, null, timePoint));
+    }
+
+
+    private void mockToolManager() {
+        doReturn(tool).when(toolManager).loadByNameOrId(eq(IMAGE));
+        doReturn(tool).when(toolManager).resolveSymlinks(eq(IMAGE));
+        doReturn(Optional.empty()).when(toolManager).loadToolVersionScan(eq(tool.getId()), eq(null));
+    }
+
+    private void mockCloudRegionManager() {
+        doReturn(defaultAwsRegion).when(cloudRegionManager).load(eq(ID));
+        doReturn(defaultAwsRegion).when(cloudRegionManager).loadDefaultRegion();
+        doReturn(nonAllowedAwsRegion).when(cloudRegionManager).load(eq(ID_2));
+        doReturn(notDefaultAwsRegion).when(cloudRegionManager).load(eq(ID_3));
+    }
+
+    private void mockPipelineConfigurationManager() {
+        doReturn(configuration).when(pipelineConfigurationManager).getPipelineConfiguration(any());
+        doReturn(configuration).when(pipelineConfigurationManager).getPipelineConfiguration(any(), any());
+        doReturn(new PipelineConfiguration()).when(pipelineConfigurationManager).getConfigurationForTool(any(), any());
+    }
+
+    private void mockPreferenceManager() {
+        doReturn(CLUSTER_DOCKER_EXTRA_MULTI.getDefaultValue())
+                .when(preferenceManager).getPreference(eq(CLUSTER_DOCKER_EXTRA_MULTI));
+        doReturn(CLUSTER_INSTANCE_HDD_EXTRA_MULTI.getDefaultValue())
+                .when(preferenceManager).getPreference(eq(CLUSTER_INSTANCE_HDD_EXTRA_MULTI));
+        doReturn(CLUSTER_SPOT.getDefaultValue()).when(preferenceManager).getPreference(eq(CLUSTER_SPOT));
+        doReturn(COMMIT_TIMEOUT.getDefaultValue()).when(preferenceManager).getPreference(eq(COMMIT_TIMEOUT));
+    }
+
+    private void mockInstanceOfferManager() {
+        doReturn(true).when(instanceOfferManager).isInstanceAllowed(anyString(), eq(ID), eq(true));
+        doReturn(true).when(instanceOfferManager).isInstanceAllowed(anyString(), eq(ID), eq(false));
+        doReturn(true).when(instanceOfferManager).isToolInstanceAllowed(anyString(), any(), eq(ID), eq(true));
+        doReturn(true).when(instanceOfferManager).isToolInstanceAllowed(anyString(), any(), eq(ID), eq(false));
+        doReturn(false).when(instanceOfferManager).isInstanceAllowed(anyString(), eq(ID_2), eq(true));
+        doReturn(true).when(instanceOfferManager).isInstanceAllowed(anyString(), eq(ID_3), eq(false));
+        doReturn(true).when(instanceOfferManager).isPriceTypeAllowed(anyString(), any(), anyBoolean());
+        doReturn(BehaviorSubject.create()).when(instanceOfferManager).getAllInstanceTypesObservable();
+        doReturn(price).when(instanceOfferManager)
+                .getInstanceEstimatedPrice(anyString(), anyInt(), anyBoolean(), anyLong());
+    }
+
+    private void mockParentRun() {
+        doReturn(parentRun).when(pipelineRunDao).loadPipelineRun(eq(PARENT_RUN_ID));
+        doReturn(false).when(permissionManager).isRunSshAllowed(eq(parentRun));
+    }
+
     private void launchTool(final PipelineConfiguration configuration, final String instanceType) {
         launchPipeline(configuration, null, instanceType, null);
     }
@@ -384,5 +553,9 @@ public class PipelineRunManagerLaunchTest {
                                        final String instanceType, final Long parentRunId) {
         return pipelineRunManager.launchPipeline(configuration, pipeline, null, instanceType, null,
                 null, null, parentRunId, null, null, null);
+    }
+
+    private ContextualPreferenceExternalResource getResource() {
+        return new ContextualPreferenceExternalResource(TOOL, tool.getId().toString());
     }
 }
